adicao dos testes unitarios na controller, cada componente com 3 tipos de teste. Atualização da Product Controller nos tratamentos de erro.
